(PC-27682)[PRO] refactor: create VenueCreationForm and VenueEditionForm folders
