+++ pro/src/components/Address/__specs__/Address.spec.tsx
@@ -5,7 +5,7 @@ import React from 'react'
 import * as yup from 'yup'
 
 import { apiAdresse } from 'apiClient/adresse'
-import { VenueFormValues } from 'components/VenueForm'
+import { VenueFormValues } from 'components/VenueCreationForm'
 import { SubmitButton } from 'ui-kit'
 
 import { AddressSelect, validationSchema as addressValidationSchema } from '../'
