+++ pro/src/components/VenueCreationForm/Accessibility/__specs__/Accessibility.spec.tsx
@@ -4,7 +4,7 @@ import { Form, Formik } from 'formik'
 import React from 'react'
 import * as yup from 'yup'
 
-import { VenueFormValues } from 'components/VenueForm'
+import { VenueFormValues } from 'components/VenueCreationForm'
 import { AccessiblityEnum } from 'core/shared'
 import { SubmitButton } from 'ui-kit'
