+++ pro/src/components/VenueCreationForm/ImageUploaderVenue/__specs__/ImageUploaderVenue.tracker.spec.tsx
