+++ pro/src/components/VenueCreationForm/Informations/SiretOrCommentFields/SiretOrCommentFields.tsx
@@ -4,7 +4,7 @@ import React, { useEffect, useState } from 'react'
 import { getAdressDataAdapter } from 'components/Address/adapter'
 import { handleAddressSelect } from 'components/Address/Address'
 import FormLayout from 'components/FormLayout'
-import { VenueFormValues } from 'components/VenueForm/types'
+import { VenueFormValues } from 'components/VenueCreationForm/types'
 import getSiretData from 'core/Venue/adapters/getSiretDataAdapter'
 import { humanizeSiret, unhumanizeSiret } from 'core/Venue/utils'
 import { InfoBox, TextArea, TextInput } from 'ui-kit'
