+++ pro/src/pages/VenueEdition/VenueEditionHeader.tsx
@@ -5,7 +5,7 @@ import { ImageUploader, UploadImageValues } from 'components/ImageUploader'
 import { ButtonImageEdit } from 'components/ImageUploader/ButtonImageEdit'
 import { OnImageUploadArgs } from 'components/ImageUploader/ButtonImageEdit/ModalImageEdit/ModalImageEdit'
 import { UploaderModeEnum } from 'components/ImageUploader/types'
-import { buildInitialValues } from 'components/VenueForm/ImageUploaderVenue/ImageUploaderVenue'
+import { buildInitialValues } from 'components/VenueCreationForm/ImageUploaderVenue/ImageUploaderVenue'
 import { Events } from 'core/FirebaseEvents/constants'
 import { SelectOption } from 'custom_types/form'
 import useActiveFeature from 'hooks/useActiveFeature'
