+++ pro/src/pages/VenueSettings/VenueSettingsForm.tsx
@@ -8,12 +8,12 @@ import {
 } from 'apiClient/v1'
 import { AddressSelect } from 'components/Address'
 import FormLayout from 'components/FormLayout'
-import { VenueFormValues } from 'components/VenueForm'
-import BankAccountInfos from 'components/VenueForm//BankAccountInfos/BankAccountInfos'
-import { OffersSynchronization } from 'components/VenueForm//OffersSynchronization'
-import { VenueFormActionBar } from 'components/VenueForm//VenueFormActionBar'
-import { WithdrawalDetails } from 'components/VenueForm//WithdrawalDetails'
-import SiretOrCommentFields from 'components/VenueForm/Informations/SiretOrCommentFields'
+import { VenueFormValues } from 'components/VenueCreationForm'
+import BankAccountInfos from 'components/VenueCreationForm/BankAccountInfos/BankAccountInfos'
+import SiretOrCommentFields from 'components/VenueCreationForm/Informations/SiretOrCommentFields'
+import { OffersSynchronization } from 'components/VenueCreationForm/OffersSynchronization'
+import { VenueFormActionBar } from 'components/VenueCreationForm/VenueFormActionBar'
+import { WithdrawalDetails } from 'components/VenueCreationForm/WithdrawalDetails'
 import { Providers } from 'core/Venue/types'
 import { SelectOption } from 'custom_types/form'
 import { useScrollToFirstErrorAfterSubmit } from 'hooks'
