+++ pro/src/pages/VenueSettings/VenueSettingsScreen.tsx
@@ -10,9 +10,9 @@ import {
   GetVenueResponseModel,
 } from 'apiClient/v1'
 import ConfirmDialog from 'components/Dialog/ConfirmDialog'
-import { VenueFormValues } from 'components/VenueForm'
-import { generateSiretValidationSchema } from 'components/VenueForm/Informations/SiretOrCommentFields'
-import { validationSchema } from 'components/VenueForm/validationSchema'
+import { VenueFormValues } from 'components/VenueCreationForm'
+import { generateSiretValidationSchema } from 'components/VenueCreationForm/Informations/SiretOrCommentFields'
+import { validationSchema } from 'components/VenueCreationForm/validationSchema'
 import { Events } from 'core/FirebaseEvents/constants'
 import { PATCH_SUCCESS_MESSAGE } from 'core/shared'
 import { Providers } from 'core/Venue/types'
