+++ pro/src/screens/VenueForm/VenueEditionFormScreen.tsx
@@ -7,10 +7,10 @@ import { isErrorAPIError, serializeApiErrors } from 'apiClient/helpers'
 import { GetOffererResponseModel, GetVenueResponseModel } from 'apiClient/v1'
 import Callout from 'components/Callout/Callout'
 import ConfirmDialog from 'components/Dialog/ConfirmDialog'
-import { VenueFormValues } from 'components/VenueForm'
-import { generateSiretValidationSchema } from 'components/VenueForm/Informations/SiretOrCommentFields'
-import { validationSchema } from 'components/VenueForm/validationSchema'
-import { VenueEditionForm } from 'components/VenueForm/VenueEditionForm'
+import { VenueFormValues } from 'components/VenueCreationForm'
+import { generateSiretValidationSchema } from 'components/VenueCreationForm/Informations/SiretOrCommentFields'
+import { validationSchema } from 'components/VenueCreationForm/validationSchema'
+import { VenueEditionForm } from 'components/VenueEditionForm/VenueEditionForm'
 import { Events } from 'core/FirebaseEvents/constants'
 import { PATCH_SUCCESS_MESSAGE } from 'core/shared'
 import { SelectOption } from 'custom_types/form'
